python/requirements.txt
fastapi
python-dotenv
requests
pyjwt
jinja2
python-multipart
grpcio-tools
cryptography
uvicorn
nervatura